Support tmux config pre 2.9
Wish there was an easier way to conditionally load tmux config options
but after spend a significant amount of time trying to support 3
variations of config options across 2.4 2.8 2.9 it did not seem to
be worth the configuration headache. This configuration will remain
so that if I have to leverage older tmux versions I can at least have
some of the creature comforts.