Tmux default to non login shell
There are a couple of things that this should address. This originally
came up because my NixOS laptop didnt seem to be loading my .bashrc
for a new shell.

But additionally this was thought to fix the duplicate pathes that get
added to my shell on OSX when I launch into a tmux session. But after
testing it doesnt seem to improve this...Regardless Id still like this
change.

Additionally, I tried to "upgrade" to tmux-256color but FreeBSD was
complaining with being unable to find a valid term and was defaulting
basic to ANSI. Anyway this term seems to find with tmux 2.9+ but making
a note so I dont try to do this again later.