Make sure ANSI chars are available to shell
While using my NixOS laptop it seems that out of the box the ANSI chars
where showing up in my "git diff" with my default pager set to less. It
seems like I can force less to respect colors with this flag but not
sure of the other side effects. Going to add this to the pager env var
to be conservative.