fixing i3 conflict of ws5 and percent
For a good while I was unable to move a window in i3 to ws5 via the key
combo: mod + shift + 5. Josh was nice enough to investigate this with me
and pointed out that shift + 5 maps to percent. And wouldnt you know it
thats what I had configured for setting the split orientation for new
windows.

Opting to split now with the | in i3 so that ws5 no longer has a
conflict.

Thanks Josh for the help!