simple and fix i3lock w/dpms
previous i3locking code was reenabling dpms since it was blanking the
screen upon locking. Added a trap to make sure to zero out (effectively
disabling dpms).

i3lock.sh was writing unnecessary tmp files and blurring was slow. New
changes make this much snappier and make the blurring more extreme.

In the future it would probably be better to read in
the existing settings of dpms and restore them after login.